+++ lib/osm.dart
@@ -5,6 +5,7 @@ export 'src/area.dart';
 export 'src/bounds.dart';
 export 'src/element.dart';
 export 'src/filter.dart';
+export 'src/pbf/apply.dart';
 export 'src/pbf/exception.dart';
 export 'src/pbf/file.dart';
 export 'src/pbf/header.dart';